fix bug where additional user with same client id would be created upon first twitch login

+++ scripts/copy-if-not-exists.js
@@ -1,3 +1,5 @@
 const fs = require('fs')
 const [from, to] = process.argv.slice(2)
-fs.copyFileSync(from, to)
+if (!fs.existsSync(to)) {
+  fs.copyFileSync(from, to)
+}
